Users can now follow and unfollow on any profile page.
Users can follow/unfollow on any profile page. The follower changes are fixed for the active user and changes to UI have been updated to only occur when on your own profile page. EX: Unfollowing removes entry from the list only on your own page now. Buttons are shown in correspondence to the active users followers and not the profile page users