[RSO] live fix (#388)
* [RSO] Create RSO sign in endpoint

Added an endpoint in auth.js with a link to sign in with Riot account

* [RSO] Create RSO oauth callback endpoint

Created /rso/oauth in auth.js which takes in access code in its URL as a query string and exchanges this code for tokens using fetch to Riot's /token endpoint.

Verified this worked using Postman.

* [RSO] Oauth cleanup

* [RSO] Add RSO Sign in button

Added button inside Header.tsx that redirects user
to /rso/signin endpoint. Modified /rso/signin to
redirect to link instead of returning a Sign In
button that would redirect to Riot sign in page.

* [RSO] Store tokens in cookies

* [RSO] Start /account/me GET endpoint

Created an endpoint in auth.js that will fetch a GET to Riot's
/account endpoint to know who logged in using RSO and get the
player's userid, player tag, and player name.

* [RSO] Add get user info route, frontend fix

Finished /rso/getUserInfo route so that it takes an access token in
the params and fetches to Riot's /accounts/v1/me endpoint to get
player username, tagline, and uuid. Added functionality to Header.tsx
to show the user's username and tagline if they logged in via RSO.

* [RSO] live fix

* Prettify Code: src/api/routes/auth.js
src/web/src/Components/global/Header.tsx
src/web/src/Pages/game/Valorant/Valorant.tsx

---------

Co-authored-by: GitHub Action <action@github.com>