[RSO] Add getPuuid endpoint
Created new endpoint getPuuid to get the correct puuid for a Riot
account. This is needed because when using the puuid returned from the
getUserInfo endpoint for any Riot API that uses the puuid as part of the
url param, response returns an error saying couldn't decrypt the puuid.
The new endpoint getPuuid should fix this by returning a usable puuid.